extract Team enum
