+++ client/schemas/common.ts
@@ -1,0 +1,8 @@
+export enum Team {
+  Philadelphia = 'Philadelphia',
+  Boston = 'Boston',
+  NewYork = 'New York City',
+  SanFrancisco = 'San Francisco',
+  SummerSF = 'Summer SF',
+  SummerNYC = 'Summer NYC',
+}
